Apply explicit casts when printing dim3 variables (ported from Balint's HIP PR)

+++ lib/tune.cpp
@@ -792,8 +792,11 @@ namespace quda
           tunable.checkLaunchParam(param);
           if (verbosity >= QUDA_DEBUG_VERBOSE) {
             printfQuda("About to call tunable.apply block=(%d,%d,%d) grid=(%d,%d,%d) shared_bytes=%d aux=(%d,%d,%d)\n",
-                       param.block.x, param.block.y, param.block.z, param.grid.x, param.grid.y, param.grid.z,
-                       param.shared_bytes, param.aux.x, param.aux.y, param.aux.z);
+                       static_cast<int>(param.block.x), static_cast<int>(param.block.y),
+                       static_cast<int>(param.block.z), static_cast<int>(param.grid.x),
+                       static_cast<int>(param.grid.y), static_cast<int>(param.grid.z),
+                       static_cast<int>(param.shared_bytes), static_cast<int>(param.aux.x),
+                       static_cast<int>(param.aux.y), static_cast<int>(param.aux.z));
           }
 
           tunable.apply(stream); // do initial call in case we need to jit compile for these parameters or if policy tuning
